Use g_timeout instead of gtk_timeout. Don't release already released source.
In some of our timer callbacks, we return FALSE/G_SOURCE_REMOVE to
trigger autoremoval of the source. In those cases, we must also remove
the source id, so we don't try to remove it again. It may have been
reused, in which case we would remove something unintended.

This fixes the runtime warnings telling us about this issue.