task properties: use GtkComboBox instead of GtkOptionMenu
GtkOptionMenu has been deprecated since forever. This
replaces it with GtkComboBox in props-task.c.

The existing code attached the numeric values to the menu
items. In the *save* code path, it grabbed the numeric
value from the current selection. On the other hand, in
the *load* code path, it assumed that the menu items appeared
in a certain order.

The new code uses the select-list GtkListStore specialization.