select-list: New concept for text/value combo boxes
A "select-list" is a specialized GtkListStore holding
a (display) text in the first column and some arbitrary
integer identifying key in the second column.

It can be used with GtkComboBox to display user-friendly
strings, while allowing the program to directly set and
retrieve a numerical constant (e.g. an enum value).

Several convenience methods that operate directly on a
GtkComboBox are offered.