gtt-date-edit.c: Remove use of deprecated GtkOptionMenu
This widget used a GtkOptionMenu to open a menu with submenus, filled
with even 15-minute intervals. The GtkOptionMenu was not actually used
to display the current value in any way - it acted only as a trigger
button to open the menu. The current/selected value was displayed in a
separate text entry.

The GtkOptionMenu is now replaced with a simple GtkButton with an icon.