menucmd.c: Prevent infinite recursion in menu_toogle_timer()
menu_toogle_timer() now considers the current (new) state of the check
menu item as the desired state, and will only trigger a timer state
change if the current state is wrong. This makes it a no-op if the state
is already correct, preventing infinite recursion when the handler is
called in response to code setting the menu item state.

Fixes #43.